backend/requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.7
django-allauth==0.63.6
django-cors-headers==4.4.0
django-import-export==4.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
import-export==0.3.1
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1